Add test for value conversion; fix bug in String2VpiValue().
Testing revealed that the parsing of a real value was actually
parsing an integer instead. Fixed (atol() -> atof()).

Also IWYU: vpi_uhdm.h was missing the vector header.

+++ templates/vpi_user.cpp
@@ -75,7 +75,7 @@ s_vpi_value* String2VpiValue(const std::string& s) {
   } else if (strstr(scopy.c_str(), "REAL:")) {
     scopy.erase(0,5);
     val->format = vpiRealVal;
-    val->value.real = atol(scopy.c_str());
+    val->value.real = atof(scopy.c_str());
   }
   return val;
 }
